requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==5.1.1
Django==5.1.1
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
moviepy==2.1.1
numpy==2.1.3
pillow==10.4.0
proglog==0.1.10
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.6.0
sqlparse==0.5.1
stripe==11.2.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
